backend/requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==3.7.1
bcrypt==4.3.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cryptography==45.0.6
ecdsa==0.19.1
fastapi==0.104.1
h11==0.16.0
httptools==0.6.4
idna==3.10
iniconfig==2.1.0
numpy==1.24.3
opencv-python==4.8.1.78
packaging==25.0
passlib==1.7.4
Pillow==10.1.0
pluggy==1.6.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
Pygments==2.19.2
pytest==8.4.1
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.2
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.24.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
